server/requirements.txt
fastapi[all]
uvicorn
opencv-python
jinja2
python-multipart
pyzbar
qrcode
firebase_admin
dotenv
google-cloud-firestore